Use size_t as data type for indices/sizes in C++ standalone

+++ brian2/devices/cpp_standalone/templates/common_group.cpp
@@ -41,18 +41,18 @@ void _run_{{codeobj_name}}()
 	{% block maincode %}
 	//// MAIN CODE ////////////
 	// scalar code
-	const int _vectorisation_idx = -1;
+	const size_t _vectorisation_idx = -1;
 	{{scalar_code|autoindent}}
 
     {# N is a constant in most cases (NeuronGroup, etc.), but a scalar array for
        synapses, we therefore have to take care to get its value in the right
        way. #}
-	const int _N = {{constant_or_scalar('N', variables['N'])}};
+	const size_t _N = {{constant_or_scalar('N', variables['N'])}};
 	{{openmp_pragma('parallel-static')}}
-	for(int _idx=0; _idx<_N; _idx++)
+	for(size_t _idx=0; _idx<_N; _idx++)
 	{
 	    // vector code
-		const int _vectorisation_idx = _idx;
+		const size_t _vectorisation_idx = _idx;
 		{% block maincode_inner %}
         {{vector_code|autoindent}}
 		{% endblock %}
