+++ brian2/devices/cpp_standalone/templates/network.cpp
@@ -66,7 +66,7 @@ void Network::run(const double duration, void (*report_func)(const double, const
     {
         t = clock->t[0];
 
-        for(int i=0; i<objects.size(); i++)
+        for(size_t i=0; i<objects.size(); i++)
         {
             if (report_func)
             {
